+++ libs/figma-tasks/src/executors/index.ts
@@ -1,2 +1,0 @@
-export { syncIcons } from './sync-icons/sync-icons';
-export { syncIllustrations } from './sync-illustrations/sync-illustrations';