+++ libs/figma-tasks/src/index.ts
@@ -4,5 +4,3 @@ updateNotifier({
   pkg: require('../package.json') as Package,
   shouldNotifyInNpmScript: true,
 }).notify();
-
-export * from './executors';