+++ libs/script-utils/src/fs/getAbsolutePath.ts
@@ -1,13 +1,21 @@
+import path from 'node:path';
 import { Task } from '@cbhq/mono-tasks';
 
 export function getAbsolutePath<T, PathValue extends string | undefined>(
-  task: Task<T>,
+  task: Partial<Task<T>>,
   pathValue: PathValue,
 ): PathValue {
   if (pathValue) {
+    if (typeof task.projectRoot === 'string' && typeof task.workspace?.root === 'string') {
+      const absolutePath = pathValue.startsWith('./')
+        ? path.resolve(task.projectRoot, pathValue)
+        : path.resolve(task.workspace?.root as unknown as string, pathValue);
+      return absolutePath as any;
+    }
+
     const absolutePath = pathValue.startsWith('./')
-      ? task.projectRoot.append(pathValue).toString()
-      : task.workspace.root.append(pathValue).toString();
+      ? task.projectRoot?.append(pathValue).toString()
+      : task.workspace?.root.append(pathValue).toString();
 
     return absolutePath as PathValue;
   }